+++ src/displaylist.h
@@ -102,4 +102,4 @@ void loadAssetMetadataByIndex(
     s32 sub_index
 );
 
-void func_80067EDC_68ADC(u16, loadAssetMetadata_arg *);
+void func_80067EDC_68ADC(s32, loadAssetMetadata_arg *);
